// src/data/model/performance/list/PerformanceItemDto.ts
export interface PerformanceItemDto {
  id?: string;           // mt20id
  name?: string;         // prfnm
  startDate?: string;    // prfpdfrom
  endDate?: string;      // prfpdto
  facilityName?: string; // fcltynm
  posterUrl?: string;    // poster
  area?: string;         // area
  genre?: string;        // genrenm
  openRun?: string;      // openrun
  state?: string;        // prfstate
  awards?: string;       // awards
}